+++ src/matrices/m4_inverse.c
@@ -6,7 +6,7 @@
 /*   By: abessa-m <abessa-m@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/08/03 15:07:28 by abessa-m          #+#    #+#             */
-/*   Updated: 2025/08/03 17:05:50 by abessa-m         ###   ########.fr       */
+/*   Updated: 2025/08/06 21:12:36 by abessa-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ t_m4	m4_inverse(t_m4 m4)
 	det = m4_determinant(m4);
 	while (i.y < size)
 	{
-		ret.m[i.y][i.x] = m4_cofactor(m4, i) / det;
+		ret.m[(int)i.y][(int)i.x] = m4_cofactor(m4, i) / det;
 		i.x++;
 		if (i.x >= size)
 		{
